Add .gitignore so it is clearer what is generated by the build process
Updated README with instructions on what it took for me to get this
to build from source and run its self tests.

By putting the src/.libs directory on the LD_LIBRARY_PATH env var, and the
samples directory on PATH, one can run the sample programs at the command line.