Schema locations on command line, negative tests
This was painful.

You have to mine the targetNamespace out of
the schema file, because Xerces C simply doesn't have
a behavior where it forces a schema onto a file.

Enhanced built-in tests with negative tests

It now has tests for SAX2Count that show it
detects invalid XML documents given
schemas provided on the command line.